fix color rgb/rgba/hex annotation
NOTE: rgba is not supported by the html-input-element, but it is still supported internally, though setting a value results in a fallback to a rgb value. If somewhere in the future a component is integrated, that supports setting rgba values, the supporting functionality is already implemented. If the value is only handled internally and the property is hidden/final/readonly, rgba is supported to be used.